cuda: Rename and duplicate kernels so that each kernel has consistent workload
This makes it easier to use NV profiling tools. For now we don't do any
extra fusion although there's certainly plenty of opportunities, but the
point here is to get rid of "generic" matmul kernels.

Note: all previous measurements of bandwidth in commits that said ~125
GB/s should really say "125 or 150 GB/s"... for some reason inference
perf has a bimodal distribution on my 4090, so the numbers should be
treated with a grain of salt. Until the source of discrepancy is
identified we'll assume we run at ~150 GB/s, but we will have occasional
issues with measurements that don't align with this observed perf.