cuda: Fuse qkv and rope kernels
For rope, we're mostly looking to get rid of an extra bubble and
processing q & k in parallel doesn't gain us any performance. For qkv,
we're looking to increase parallelism so we need to process q, k and v
at the same time which requires a little bit of indexing adjustments as
kv outputs are shorter.