Run one warmup token before generation
The first token seems to take a little bit of an extra time for CUDA
kernel loading (and, worst case, compilation); even in the best case it
takes ~8ms extra which throws off the timing enough to be noticeable.

We now run one forward pass for warmup, which fixes this issue and also
doubles as a way to preload weights when CPU backend is used.