Skip logits computation and skip last layer update in prompt mode
When logits are unused (since we're evaluating prompt tokens), we can
skip the classifier multiplication, skip attention computation and FFN
in the last layer, and remove the device sync.

This marginally accelerates prompt processing; in the future we'll
probably run the prompt in batch mode, but this change is very simple so
it seems reasonable to do it now.