Switch to warp-parallel attention mixing
With this change we perform the same adjustment to attention-mixing -
warp-parallel summation with kv indexing split between blocks and
threads similar to attn_score change.

To make this change work well wrt performance, we need to transpose the
value in the kvcache so that all positions are stored contiguously for
each head, since this maximizes the locality during mixing. Fortunately
it's trivial to transpose during kvcache update and this doesn't cost
anything.

This makes attn_mix between 4.5x (llama) and 9x (mistral) faster.