Rename all tensors
Initially the hope was to be able to use arbitrary models from HF
directly, but due to the need to store tokenizer data and
hyperparameters we really need model conversion anyway.

With this change we now explicitly convert a set of expected layers in
convert.py; this helps ensure that model files stay correct (although we
don't validate shapes in convert.py atm), and improves the naming to
closer match between models and C code.