Switch rmsnorm weights to float
This doesn't have a tangible impact on precision, but this will make
future conversions of actual weights a little easier to do mechanically
as we don't want these to use fp8 et al.

This costs us ~20us/tok (out of ~16ms), which is something that may need
to be reconsidered later.