Rewrite token merging using a binary heap
Instead of an N^2 algorithm that repeatedly rescans pairs, do the
optimal thing which is to maintain a heap of merge pairs and merge them
in order of maximum score.

When merging a pair, existing pairs get invalidated; to avoid doing any
explicit work on invalidation, we detect this by comparing if the token
at the original index is in tact.

To avoid moving tokens after every merge, we fill holes with -1 and
remove them in a linear compaction pass at the end.

Merging two tokens deletes one merge and creates up to two new ones;
this results in a worst case heap size of tokens*2 which is easy to
preallocate up front.

With this optimization we no longer need tok/s measurement on
tokenization as it's too fast :)