Update README.md
Add notes about e4m3. Even though the exponent range is different from e5m2, scaling all tensors to align the maximum value with the upper range of the type results in significant improvements in perplexity. Unfortunately, prior to SM9.0, fp8 conversions to float are emulated, and this is particularly expensive for e4m3 since the exponent range is different. Hand-coded conversions perform better than what CUDA headers do by default, but the performance still drops appreciably (~865 GB/s for e5m2 => ~750 GB/s for e4m3). It might make sense to reconsider this for the future generations of GPUs -- since this requires tensor scaling, for now the code to work with e4m3 is omitted.