Micro-optimize rmsnorm with dynamic shared memory
Instead of reading from x[] twice we save the value (scaled by weight)
in the first loop to shared memory which is dynamically allocated based
on dim. We only run one block anyway so there's no occupancy impact, but
this makes rmsnorm a tiny bit faster (5us => 4.2us per call which adds
up to ~50us savings on Mistral).

For some reason it's important to prescale by weight, as leaving the
weight scaling in the second loop is slower.