Generalize support for attention sinks to multiple tokens
The paper recommends using 4, but based on some limited testing 2 might
be sufficient - 1 is not enough in some cases. For now just hardcode 2.

In the future it might make sense to move this control to the caller of
forward() to share the code between CPU & CUDA, but for now it's
probably good enough as it is.