tools: Try to ensure output tensors are 256b-aligned
This doesn't really matter for CUDA, but may be valuable for CPU
implementation. Unfortunately safetensors doesn't allow for padding
between tensors, so our only recourse (short of including fake padding
tensors which I'd rather not do for now) is to add tokenizer tensors at
the end and hope that all tensors are individually aligned, which they
seem to be for all models we support.