tools: Initial implementation of cudaprof
cudaprof is a more general version of the profiler which injects into
the target application and gathers kernel activity via CUPTI.

One downside here is that we don't have easy access to the estimated
bandwidth numbers. Also this is a separate binary that needs to be
built, although it's easy enough and the shell script makes this
painless. The upside is that we get access to more precise kernel
execution times than possible via events, and in theory we could get
future access to hardware metrics via the same API.