Use #pragma omp simd in dotprod "fallback" code
Clang does a really good job for both of these loops, and with the
pragma it generates code that runs at peak performance on both x64 and
ARM64. Unfortunately, gcc doesn't vectorize either of the loops well and
doesn't seem to recognize the simdlen pragma, so for now we'll keep the
code with intrinsics.