Implement initial scaffolding for fused attention kernel
The kernel itself is hyper inefficient right now as it's effectively
single-threaded per head, but this lets us set up the external code to
call the kernel and iteratively optimize from here. The hope is that we
can get faster attention for small sequences using fused kernels, and
maybe use separate kernels when kv_len gets large.

The kernel is only used when CALM_FMHA env variable is set.