Fix CALM_POSO=x handling during warmup forward call
In the future we should probably remove warmup forward altogether, but
the biggest current issue is that it poisons the statistics when
CALM_POSO is specified and forces us to wait for all layers to finish
when using Nsight Compute. Both are fixed by using the same offset
during warmup call.