Improve attn_score performance by removing warp-parallel dp
Warp-parallel dot product is a bad idea for short vectors. For Phi2,
head_size is 80; for other models it's generally 128. Because we compute
two elements per loop iteration (before unrolling), we end up wasting
warp lanes on Phi2 and only doing two iterations for other models; after
this, the overhead of warp reduction and other computations dominates.

Instead we switch to a somewhat traditional column wise computation...
except that to load 4b per lane, we need to transpose the keys such that
two values are available for each timestep. We could also use something
more similar to attn_mix but this works for now.

This has a small penalty on short context lengths unfortunately, but on
long context lengths it recovers the pace and ends up producing
relatively consistent overall model bandwidth on a variety of context
sizes.