Fused attention kernel was a mistake so remove it
It's one more kernel to think about, and it needs to adapt to
optimizations elsewhere due to data layout dependencies, which makes the
implementation less composable. When kv_len is large, it will be
impossible to reach satisfactory performance with a fused kernel due to
softmax dependencies. Flash Decoding work recognizes this and has two
kernels - one computes flash attention using a number of K-splits, and
another one combines the results. We have three kernels but we have an
opportunity to get rid of the middle softmax, and the current structure
is easier to optimize than a fused monster.