Implement ~flash decoding
This is not quite flash decoding, but it uses the same softmax tiling
trick with logsumexp being stashed and then using these to rescale the
results. It's fairly convenient to do from the attn_score kernel since
each warp processes 32 elements at a time, so we might as well split
into 32-sized groups.

Unfortunately, logsumexp need to be softmaxed themselves using max/sum
for numerical stability (critical for Phi); because of this, whatever we
gain on removing a smaller sync kernel between score and mix, we lose on
redundant work in mix. As a result, performance is approximately
neutral, and this is after having to reuse work through shared memory :(

I don't think this should really be in the code long term but just in case
I'm wrong I'd like to preserve this for posterity. The next commit will
back this out as the perf gains don't justify the effort.