Replace model bandwidth estimation with tensor sizes
This makes it easier to count the model size in both params and bytes
without relying on the specific model architecture. The model bandwidth
number is currently assumed to be the same as model size; this will
change when we implement MoE but should be reasonably easy to adjust
for, given uniform tensor names.