Introduce initial version of gf4 quantization
gf4 stands for "group float 4bpw" and quantizes 8 floats into 32 bits by
using an fp8 scale per group and a 3-bit quantized signed integer. The
signed integer has [-4, 3] range; we use this to keep the scale signed
such that the maximum value gets converted perfectly, at the cost of
lost range in the opposite sign. This helps for groups with outliers.

There's freedom during quantization to pick an arbitrary scale value. In
the future this can be used to push quality further without changing the
format. There might also be reasons to use non-uniform values for the
quantized integer, or use a different dequantization formula (currently
assumption is that the int is dequantized via q/4). The format is not
set in stone and may evolve in the future.

The small group size will hopefully compensate a little for limited bits
on various components; this format was chosen to keep group size aligned
in terms of both the number of weights (8) and number of bytes (4) for
efficient inference. Wider formats (eg 16 weights in 8 bytes) would be
possible and might be preferable in certain cases.