Unroll loop in gf4 matmul_warp overload
Normally the backend compiler does it for us but in this case the loop
body is too large. Unrolling here at least 2x is critical to get good
memory throughput.

With this we require the vector width to be divisible by 512 which
happens to work with all models we support but is definitely at the
limit of acceptable alignment. There are probably other loop structures
that would work here in the future though, and we can always add a
smaller loop stride back if need be.