tools: Accelerate quantization by running it via CUDA
This gets a substantial speedup (~7x) over CPU backend. It's mostly for
the wrong reasons - CPU code takes ~40 seconds to process ~14 GB of
input data, and a non-sloppy C code would do this in under a second -
but for now we'll make do with ~6 seconds to process a 7B model.