tools: A few more tweaks to convert/gf4
These mostly intend to address the performance issues during conversion
when not using CUDA (in case we need to do this again), and get us to
~30s (from ~40s) by making sure we use types that are as narrow as
possible and allocate a little fewer tensors without impacting code
readability.

gmax is kept in its original precision; float() was necessary to make
sure "+ 4" computation has enough precision, but float16 is good enough
for that, and division can apparently stay in bfloat16 for now too. None
of these tweaks affect perplexity on Mistral 7B.