Optimize AVX2 dotprod_gf4 variant
We decode 4 groups at a time, which allows us to slightly increase ILP
as well as to do fp8->half->float 4 values at a time. Additionally, we
replace the math involved in converting the integers with a precomputed
table, which is a little faster as it fits AVX2 vpermps perfectly.

We reach ~42 GB/s give or take on Zen4 now which is close to bandwidth
peak. This requires fewer threads than before as we're reasonably
optimal wrt compute density.