tools: Work around slow cold FS loads via mmap->cudaMemcpy
cudaMemcpy from mmapped data ends up reading the file with a block size
that is not super efficient for SSDs. This isn't a problem if the file
is in FS cache, but is a problem if it's not. Loading the tensor on CPU
works around this problem (the entire conversion goes from 18 seconds
before this fix to 10 seconds with this fix, using Llama2 7B hf->gf4).

This needs to be fixed in PyTorch but for now work around the issue :-/