Avoid division by zero for almost-zero groups
When gmax is 0 (fp8), dividing by it yields NaN (if the source
weights were exactly zero) or +-Inf (if the source weights weren't zero
but were smaller than minimum representable fp8).

By itself division by zero isn't a problem - we just select an index
that may or may not make sense - but it results in out of range indices
as well as issues with further math we might want to do here, so replace
results of division by zero with 0.