Rework min-p sampling to work directly on the logits
This saves the overhead of doing expf on tokens that will not be
considered anyway, and makes the overhead of sampling fairly negligible
on reasonable minp values. For small minp (eg 0) the code is doing
roughly the same computations as before although it's skipping the
redundant normalization pass as we will need to normalize the selected
probabilities anyway.