Use warp-parallel top k expert selection for MoE
This saves us ~0.5us per layer (~15us per token), which is not much but
it's a cool technique so for now let's keep it. This is especially
beneficial if we decide to split matmul from the selection kernel (which
is generally negative on performance on 4090, but might help us on H100
and other DC grade GPUs, and on 4090 the extra costs would be subsumed
by this optimization).