Remove incorrect assertions from warppar matmuls
We do not actually require that n is divisible by the loop multiple. If
it is, we will have all threads active all the time which is great. But
even if it's not, we will just have some threads inactive in the last
slice which is automatically handled by the loop structure.

Some models like Qwen14B have sizes that aren't divisible by 256 (gf4
warp width in elements), which this change fixes the inference for.