Fuse RoPE and matmul_qkv kernels
matmul_qkv is in a fairly optimal position to produce two elements per
block - we already do that for fp8 and gf4 due to matmul_par - which
makes it easy to rotate the resulting complex value and write it to KV
cache directly. This removes the need for an extra rope kernel at a very
small extra cost to the matmul_qkv kernel which ends up as a net win of
about 1% of inference of models like Mistral and Llama.

The only fairly awkward bit here is that we still need to rotate kv_sink
tokens; to avoid extra overhead at lengths over the context size we just
do that in matmul_qkv; it might make sense to split it at some point.