Remove attn_softmax kernel by splitting it into score/mix
softmax, broadly, needs two synchronous parts: computing max value for
numerical stability, and computing exp(v-max) (and it's sum).

Instead of using a separate kernel for this, we can compute max as we
are computing scores, using warp max reduction + global atomics. Once we
have a per-head max, we can just apply the rest of softmax as we are
doing mixing. This technically redundantly recomputes expf(), but it
should not matter as much as the mix kernel should be bandwidth bound.

This ends up getting us ~1% speedup early on in the generation and being
more or less flat at the end of the context wrt performance.