tools: Add speed of light estimation notebook
Given the theoretical peak bandwidth and FLOPs of a given GPU we should
be able to estimate "ideal" inference performance. It will vary with
architecture, context length, and batch size.

In this change we build a basic config with the model structure and can
compute the size of KV cache as well as required bandwidth and FLOPs for
inferring a single batch of tokens (assuming each new token is
auto-regressively generated vs the earlier tokens, so all tokens in a
batch are new).