Check CUDA last error after synchronizing
If any kernel launch results in an error due to OOM when the kernel
can't be compiled, the error is set globally but isn't communicated to
the stream. As a result we have a "functioning" pipeline where kernels
simply don't run and the output logits are non-sensical.

Ideally we'd have all memory requirements 100% set at prepare() time but
for now let's guard against this during inference.