Automatically switch to fp8 kvcache for longer contexts
When using CUDA we now support dynamically switching between fp8 and
fp16 kvcache. We could explicitly control this but for now for
simplicity let's just pick fp8 when a longer context is specified (4096
is the limit for automatic context selection so fp8 will need to be
opted in via -c 8192 or something like this as a result of this change).

CPU only supports fp16 kvcache for now.