Defer softmax normalization until attn_mix
We really want to compute exp(t) once as exp() is fairly expensive to
compute, but we can omit the normalization until mixing. It's almost
free to do during mixing as we already need to read the entire set of
scores per head, but this slightly simplifies softmax which is more
synchronous, which helps a little bit for long contexts.