Add support for Qwen1.5
Note: the HF arch is called "Qwen2" but the model files are in Qwen1.5
repositories; maybe Qwen2 will be a larger scale retraining that will
follow? For now use "qwen2" name.

The model is much more similar in structure in HF exports to Llama; I'm
not 100% confident that we are running it correctly though (aside from
rmsnorm eps which needs to be fixed separately).