tools: Fold norm biases into other biases for Phi2
This is a little less future proof because there might exist models that
have layernorm with a bias but no MLP biases, but fundamentally the
biases here are just an oversight and it will be easier to maintain two
types of norms that don't have biases (and just have scale weights).

We can fold biases into biases of linear layers that use norm as an
input; for Phi this means folding norm biases for individual layers into
QKV and MLP1 transforms, and folding output norm bias into classifier
bias.

This change just does this as a model transform; a followup change will
remove actual support for norm bias.