tools: Add support for OLMo conversion
OLMo models are structurally almost identical to LLama, with the
exception of using layernorm instead of rmsnorm (but without
weight/bias). Since we no longer use biases for layernorm this is fairly
easy to represent; this change adds the conversion part, but it will
need changes on the inference side to select the correct normalization
algorithm as well.