Cooperative group work starts
Following the Groq launch I am now curious what the peak perf is for
tensor parallel BS=1 inference on 8xH100. The current approach is likely
a dead-end for peak inference performance in such a difficult setting
due to the cost of inter-device synchronization across CUDA streams. We
need to get much more aggressive about fusion and how we program this
hardware.

The first step on this journey is to implement a fully fused inference
using cooperative groups; the hope is that we can do this without losing
performance compared to current approach on a single device. If that
works out then we can scale out to N GPUs.

We will do this one kernel at a time and maintain the old kernel set
separately. For simplicity for now we only support LlamaLike with CG,
although in the future it would make sense to at least support Mixtral.

This commit is a first in a long series, that's just fusing MLPs.