Fuse attention mixing into coop kernel
We reuse attn_warpdot between two implementations as is. The actual
computation scheduling is the same as before (as in, consecutive warps
process multiple duplicated heads for GQA to ensure maximum cache
efficiency). For simplicity for now we do this with a flat loop which
means we end up doing a little bit of expensive integer division math -
something to look into in the future.