Fuse attention scoring as well
The loop structure here is truly bizarre which follows from the odd 2D
layout of key buffers and the parallelism we've chosen. It's likely that
this can be reconsidered with the cooperative structure but for now
we're replicating things as close as possible to the baseline.

We also need launch bounds attribute as without it the compiler doesn't
compile the scoring loop very well, presumably due to register pressure.