Move RoPE sink handling out of the RoPE kernel
When the kernels were fused, this logic wasn't fixed to actually pass
proper arguments to the kernel so it's been broken ever since.

This is great as it's an opportunity to extract it to a separate kernel.
This kernel only needs to run once per token (as it can rotate all
layers simultaneously) and it only needs to run when kv_sink is active.

Even when the kernel is active, the time it takes is insignificant (~4us
per token on 4090 for Mistral).