Support custom head_dim during inference
This is needed for Gemma model; until now we've been assuming that
head_dim = dim / n_heads, but with this change we explicitly store the
head dimension without computing it.

This change isn't fully complete because there are still some
assumptions about dim/kv relationship that we need to fix.