Use a tiled block layout in coop kernel
Instead of processing groups of 32 blocks per SM, process groups of K
blocks per SM, with groups interleaved across SMs. This reduces the tail
effect for GPUs like A100/H100 where SM number is not a power of two so
it doesn't neatly divide dimensions of matrices we work with.