Reduce rmsnorm cost in coop mode
rmsnorm does a reduction on x^2 followed by the rescaling of the output.
Instead of doing the rescaling, we can defer it to the next (parallel)
stage - doing matmul on the unscaled output.

This requires a scalar scale, so we need to premultiply the output with
the per-component scale as we are doing the reduction.