Replace the uses of RunState::xb2 with RunState::q for CUDA
xb2[] is dim-sized, which makes it (theoretically) unsuitable for
operations that expect q_dim in case q_dim > dim. In practice this works
fine because q_dim <= dim in models we support, but it is more correct
to use a q_dim sized vector.

Since q[] is unused at this point in the pipeline we might as well reuse
it (this looks a little strange from the data flow perspective, but is
correct and efficient).