Stack MoE layers instead of keeping them split
With this, each MoE layer now is a larger tensor with all experts placed
one after each other in memory. This is more convenient and performant
to do inference for (because there is no need to indirectly load the
weight address); this might be less convenient for multi-GPU scenarios
when each expert layer will probably need to be sharded across all GPUs,
but it should still be doable in theory, and this change will make coop
support for MoE more practical.

This invalidates existing .calm files but that's just a problem for
Mixtral.