Remove stride argument from matmul_warppar
We've added this a while ago in anticipation of k-split optimizations
but these never happened; additionally to do k-split you need to offset
weights but that offset is dependent on the type so the code isn't
generic anyway.

For now let's simply remove the stride support; if we need it for a
specific use case we can reimplement it properly.