Implement Mixtral support in coop kernel
For now we use a slightly different strategy for MoE gate where we
separately do a matmul on 8 different SMs and then aggregate the
results. Following the approach that the separate kernel uses regresses
performance dramatically for some reason. This needs an extra sync stage
but we'll come back and fix that.

To keep things minimal and performant we unify MoE and non-MoE
inference: non-MoE is processed assuming there's just one active expert
with a fixed index and weight. There is a little bit of extra index
computation logic but that doesn't seem to hurt performance. We might
come back to this.