Switch to a single block for moegate matmul
Similarly to the non-coop approach, it is more efficient to multiply by
moegate on one SM (it's just 8 outputs). This allows us to reduce
synchronization and ultimately results in better performance.

I've also experimented with running moegate matmul and rmsnorm in parallel
on different SMs (which is possible if we keep moe_gate_warp stage
separate as it needs access to rmsnorm), and that yields the same
performance improvement but results in slightly different outputs due to
the need to premultiply moegate by rmsnorm and that matrix being
quantized.