Fuse rmsnorm with matmul_cls for coop mode
For some reason, putting final rmsnorm into coop kernel reduces the
performance (due to compiler issues). Instead, we fuse rmsnorm into the
output kernel and restructure this a little bit to be similar to coop.

The total performance is about the same; this has a benefit of not
having to rely on xb[] which might be useful for multi-GPU later.