Keep intermediate normalized state in shared memory
Instead of doing rmsnorm on one SM and having everyone else wait for it,
we now independently do rmsnorm into shared memory in each thread
(expecting the same results as the computation is completely
deterministic). This seems wasteful but it allows us to remove a sync
point, and other SMs aren't doing anything useful during rmsnorm sync
anyway, so this ends up being a little faster.