Add support for QKV bias for coop mode
To avoid additional overhead from layer pointers we are starting to
consolidate tensors further, for now just fusing q/k/v biases into a
single long vector. Unlike matrices, this is trivial to decompose as
it's always float32, and this tensor is unusual anyhow.

This was the only thing missing for Qwen support which works in coop
mode now. Existing Qwen/Phi model files will need to be reconverted.