Implement support for Olmo model in coop mode
Olmo needs layernorm, which needs to subtract mean. Our unfused
layernorm implementation computes mean and variance in one pass but
since we rely on scalar post-scaling in fused mode, we need an extra
pass to accomodate that, so just subtract mean as we go.

This makes coop mode fully featured except for Phi that is unlikely to
be supported long term.