Implement support for MiniCPM
The model architecture is fairly standard with tied weights, but it has
a couple scaling factors that other models don't have. For now we fuse
them into the relevant matrices which seems to work just fine.

The vocab size for this model isn't divisible by 32; we might need to
pad it in the future, but for now coop mode can support it.