Switch to multi-arch SASS nvcc builds
For now we compile sm_80 and sm_90; running on 4090 will use sm_80 SASS
as there is no PTX embeded anymore so there's nothing to JIT-compile. We
may revisit this in the future to take advantage of fast fp8->half on
sm_89.

If a single-arch build is desired, CUARCH=native or CUARCH=sm_89 will
generate one.