Implement instrumented profiling for coop mode
Since our external profiler doesn't have visibility into coop stages we
need to manually collect the performance information.

To do that, we use a global timer which counts nanoseconds; clock64 does
not appear to produce consistent timings.

To simplify configuration space, we only dump the profiling information
when cudaprof is used to profile the executable.