Replace syncgrid() with syncgrids()
Instead of using coop kernel barrier, we use a separate allocated cell;
this will make it possible to use this to synchronize across multiple
GPUs (if this approach even works...).

The syncgrids() core mirrors CUDA's
cooperative_groups::details::sync_grid with the caveat that we are only
putting barrier_sync at the end. I'm not sure what the motivation is for
placing two barriers; it seems as if memory visibility doesn't change,
and we already have an implied guarantee that all threads/warps arrive
at the sync point eventually... This slightly reduces sync overhead.