Clean up and fix new grid synchronization code
Instead of __barrier_sync, we need to use __syncthreads to be compatible
with earlier versions of CUDA where the semantics seemingly was
different. Additionally, while the previous change externalized the
barrier to make it feasible to share with other devices in the system,
this strategy ended up being suboptimal based on actual multi-GPU
testing; a multi-stage sync is better in these scenarios and syncgrid()
is an important part of that, so for simplicity for now we'll assume
syncgrid() works on the internal barrier.