Score two elements for each token in parallel
Instead of every thread of a warp working on an individual element of a
sequence, we can split the sequence in multiple chunks and use shuffles
to combine the score at the end.

There's a limit to the usefulness of this technique: we used to do this
fully warp parallel and that had its own issues with efficiency. But
with this moderate change, we should still be safe wrt efficiency:
optimally, the score loop processes 8 elements per iteration (across 2
threads) * 4 unrolled iterations = 32, and every head dimension is
divisible by 32.

(note: divisibility here is purely an efficiency requirement; if it does
not hold, we will just fall back on an non-unrolled loop, losing a
little bit of perf)

This change keeps the performance on long sequences but significantly
improves performance on short sequences (1.5-2x) for scoring. It might
make sense to do another 2x later.