Optimistically skip numeric softmax optimization
If all scores are in a reasonable range (defined as "exp(x) doesn't
underflow or overflow"), we don't really need to run the softmax
correction. This is great because this means we don't need a whole
separate sync stage as we can precompute exp(x).

Unfortunately if the scores did overflow we can't easily recover them.
For now let's save the original scores *and* the exp(), and condition
the fallback on any score being outside of [-40, 40] range.

There are better ways to do this optimization, but for now this seems
serviceable.