Remove embed_scale support for now
We ended up not using it as it's better to fuse the scale into the
embedding vector, and it's possible to recover it using the final norm.

A different approach with per logit or per latent dim scale might be
useful in the future but this needs to be implemented for a specific
reason.