Always use wider loads for fp8 kernels
This is necessary for both A100 and H100, and since A100 is SM8.0 our
previous trick of hiding this behind CUDA_ARCH ifdef doesn't work.

Let's just use this uncondtionally for now. The issues with shared
memory coalescing are benign and can be resolved in the future if we
switch to half-precision activations.