Implement tokenization for special tokens
Special tokens don't follow BPE merge rules and as such need to be
discovered in the initial scanning phase; for safety for now we limit
them to <|...|> delimited sequences.

Also, for models where "eos" signals end of turn, we might need to embed
end of turn (for the query) into the initial prompt; adjust terminating
condition accordingly.