Initial scaffolding for half-precision activations
To see if increasing compute density for matmul kernels will be helpful
we need to support half-precision activations. There's a few caveats for
doing that so for now we allow forward to specialize on the activation
type, and expect individual matmul kernels to become available gradually
where performance improves.

This change by itself doesn't introduce half-precision activations yet.