Add Metal buffer upload as well as other scaffolding
run.c now uses more or less the same interface for Metal as it did for CUDA; now we "simply"
need to implement prepare/forward. In upload_metal we now upload the buffer and return the
buffer handle as pointer so that the rest of the code doesn't need to change.