Metal: Use a slightly more ALU-efficient gf4 unpack
We rearrange the bits in gf4 encoding to align groups of 4 3-bit values
in high and low 16-bit halfwords; this allows us to shift and mask the
value pair to the top 3 bits of each halfword, which can be converted to
a signed floating point value -- converting each half is just one instruction
since 16-bit halfs of each register can be addressed individually by
the conversion instruction.

This also allows us to multiply the sum by scale separately.

Overall this reduces the ALU pressure, making it a non-limiting factor.
This change by itself is just a ~1% uplift but it unlocks further optimizations
by significantly shifting the bottlenecks.