Metal: Pre-process gf4 weights in memory to make matmul a little faster
To accelerate matmul further we need a slightly different order and structure of bits, and
it's getting relatively expensive to convert this on the fly during a matmul, so we now
shuffle the gf4 weight bits in memory. This allows us (for now) to remove a XOR, and with
another small tweak wrt bit math we get a ~2% speedup. Slightly more should be possible
if we also change the bit order, although we need to be careful to keep fp8 scale fast to
decode, so that will have to happen separately.