Metal: Implement full MoE support
For now we use the same strategy as CUDA, where a non-MoE model is assumed to have
one expert and runs the same kernels. This seems to have a small performance penalty
on Mistral though (~0.5-1%?), which we'd need to fix eventually, either by specializing
the MoE kernels or by tweaking the code somehow.