Metal: Use half precision activations for attn_out x gf4
We had one last kernel that didn't use half precision activations when using gf4 weights;
it was a little bit odd because fixing that required to specialize attn_mix for gf4
weights even though mix is weight-independent. However, this tweak has no impact on quality
and it does make things ~1% or so faster on Mistral 7B gf4 so let's just do this.