Implement basic support for end-of-turn tokens
Llama 3 requires EOT token to actually terminate the output; for now
we'll just look for tokens like this by name as they aren't present in
the model metadata anyway.

Note that chat mode doesn't work with Llama 3 even with this change
because it uses [INST] wrapping instead of Llama 3 chat template.