Rework chat frame detection to use tokenizer exclusively
Almost all chat formats, with the exception of baseline llama, rely on
the presence of special tokens; as such, it would make sense to use
tokenizer instead of the model architecture to determine the chat style.

This is a little less ad-hoc and less prone to misbehavior on fine
tunes.