Remove add_dummy_prefix logic
Different models disagree on whether or not a dummy prefix is a good
idea to add. This was always a gross hack, it doesn't match instruction
tuning formats for a few models, and it makes the tokenization
irreversible which is just a terrible property to have.

Removing it makes the code cleaner and any model that suffers from it
only has itself to blame.