Compute hash for logits of the last token
This helps us ensure determinism of computation, as there are cases
where non-determinism in the processing still results in the same output
as the error is under the sampling threshold. We only compute these for
the last token as the computation is not free, and this way we minimize
the overhead -- there might be cases where only intermediate logits
change but that's fairly unlikely as the noise would usually propagate
through KV cache.