Fix non-deterministic evaluation in CUDA
It's not clear why this is necessary; CUDA's implementation of
cooperative sync does bracket the sync code in syncthreads from both
sides but it seemed as if just one syncthreads is enough. Moreover we
seem to only get non-deterministic evaluation in practice from two calls
(as in, adding syncthreads before just two syncgrids suffices), and it
must be syncthreads and can't be a threadfence.

None of this makes sense so this will likely need further investigation
into what memory ordering issues et al we are running into here, but for
now eat a tiny bit of perf for correctness.